# Conflicts:
#	ubbprotool-pc/src/main/java/com/chk/ubbprotool/ubbprotool/Model/Teacher.java
#	ubbprotool-pc/src/main/java/com/chk/ubbprotool/ubbprotool/Repository/TeacherRepository.java
#	ubbprotool-pc/src/main/java/com/chk/ubbprotool/ubbprotool/dto/CourseDTO.java
#	ubbprotool-pc/src/main/java/com/chk/ubbprotool/ubbprotool/dto/MessageDTO.java
#	ubbprotool-pc/src/main/java/com/chk/ubbprotool/ubbprotool/dto/TeacherDTO.java
#	ubbprotool-pc/src/main/java/com/chk/ubbprotool/ubbprotool/dto/UniversityClassDTO.java

+++ ubbprotool-pc/src/main/java/com/chk/ubbprotool/ubbprotool/Helper/Weeks.java
@@ -15,7 +15,7 @@ public class Weeks {
     @Id
     @GeneratedValue(strategy= GenerationType.AUTO)
     @Column(name="week_id")
-    private int weekId;
+    private Long weekId;
 
     @Column(name="week_number") @NonNull
     private int weekNumber;
