+++ ubbprotool-pc/src/main/java/com/chk/ubbprotool/ubbprotool/Service/MessageService.java
@@ -9,11 +9,11 @@ import java.util.List;
 public interface MessageService {
     void createMessage(MessageDTO message );
 
-    void deleteMessage(int messageId);
+    void deleteMessage(Long messageId);
 
     void upgradeMessage(MessageDTO message);
 
     List<MessageDTO> findAllMessages();
 
-    MessageDTO findById(int id);
+    MessageDTO findById(Long id);
 }
